+++ src/pages/dashboardPage/dashboardPage.scss
@@ -1,0 +1,9 @@
+.dashboard {
+  display: flex;
+  flex-direction: column;
+
+  @media screen and (min-width: 1024px) {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+}
