+++ src/pages/loginPage/loginPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Title from '../components/Title';
-import LoginForm from '../components/LoginForm';
+import Title from '../../components/Title';
+import LoginForm from '../../components/LoginForm';
 
 const LoginPage = () => {
   return (
