+++ src/pages/registerPage/registerPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Title from '../components/Title';
-import RegistrationForm from '../components/RegistrationForm';
+import Title from '../../components/Title';
+import RegistrationForm from '../../components/RegistrationForm';
 
 const RegisterPage = () => {
   return (
